fix(types): corregir horasVuelo Integer→Double en DTOs y MisionRestController

+++ gestion/src/main/java/com/gestion/qnt/controller/dto/PilotoResumenResponse.java
@@ -10,7 +10,7 @@ public record PilotoResumenResponse(
         String nombre,
         String apellido,
         String email,
-        Integer horasVuelo,
+        Double horasVuelo,
         Integer cantidadVuelos,
         LocalDate cmaVencimiento,
         EstadoUsuario estado,
